# sdks/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiKeyCreatedResponse": {
        "description": "Segredo `key` exibido UMA única vez, na criação.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "key": {
            "description": "Segredo completo — exibido apenas nesta resposta",
            "title": "Key",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "prefix": {
            "title": "Prefix",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "prefix",
          "key"
        ],
        "title": "ApiKeyCreatedResponse",
        "type": "object"
      },
      "ApiKeyResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_used_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "prefix": {
            "title": "Prefix",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "prefix",
          "status",
          "created_at"
        ],
        "title": "ApiKeyResponse",
        "type": "object"
      },
...
      "Body_dashboard_create_key_portal_keys_post": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "Body_dashboard_create_key_portal_keys_post",
        "type": "object"
      },
      "Body_login_submit_portal_login_post": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "Body_login_submit_portal_login_post",
        "type": "object"
      },
      "Body_signup_submit_portal_signup_post": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "Body_signup_submit_portal_signup_post",
        "type": "object"
      },
...
      "CreateApiKeyRequest": {
        "properties": {
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateApiKeyRequest",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "Modelo padrão para respostas de erro.",
        "properties": {
          "detail": {
            "description": "Descrição do erro",
            "title": "Detail",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Código interno do erro",
            "title": "Error Code"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Timestamp do erro",
            "title": "Timestamp"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "description": "API key gerada no portal self-service, enviada como 'Authorization: Bearer <key>'. Requerida pelos endpoints de dados e busca semântica.",
        "scheme": "bearer",
        "type": "http"
      },
      "SessionAuth": {
        "description": "JWT de sessão do portal (login self-service). Não usar para chamadas de API de dados/IA — use ApiKeyAuth.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API pública que expõe toda a Base Nacional Comum Curricular (BNCC) do Brasil: dados determinísticos das três etapas, acesso self-service por API keys, documentação automática e busca semântica com IA (não-oficial).",
    "title": "BNCC API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/keys": {
      "get": {
        "description": "Lista as API keys da conta autenticada (ativas e revogadas), sem expor segredo ou hash. Exige sessão do portal.",
        "operationId": "list_keys_api_v1_keys_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApiKeyResponse"
                  },
                  "title": "Response List Keys Api V1 Keys Get",
                  "type": "array"
                }
              }
            },
            "description": "Keys da conta autenticada (sem segredo/hash)."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sessão ausente, inválida ou expirada."
          }
        },
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "summary": "Listar as API keys da conta",
        "tags": [
          "API Keys"
        ]
      },
      "post": {
        "description": "Gera uma nova API key para a conta autenticada. Exige sessão do portal **com e-mail verificado** (FR-007). O segredo completo é exibido **uma única vez**, nesta resposta — apenas o prefixo é recuperável depois.",
        "operationId": "create_key_api_v1_keys_post",
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "padrao": {
                  "summary": "Key de produção",
                  "value": {
                    "name": "Minha key de produção"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyCreatedResponse"
                }
              }
            },
            "description": "Key criada; o segredo completo (`key`) só aparece nesta resposta."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sessão ausente, inválida ou expirada."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "E-mail não verificado — verifique o e-mail antes de criar keys."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "summary": "Criar uma nova API key",
        "tags": [
          "API Keys"
        ]
      }
    },
...
    "/portal/keys": {
      "post": {
        "operationId": "dashboard_create_key_portal_keys_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_dashboard_create_key_portal_keys_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dashboard Create Key"
      }
    },
    "/portal/keys/{key_id}/revoke": {
      "post": {
        "operationId": "dashboard_revoke_key_portal_keys__key_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "key_id",
            "required": true,
            "schema": {
              "title": "Key Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dashboard Revoke Key"
      }
    },
    "/portal/login": {
      "get": {
        "operationId": "login_page_portal_login_get",
        "parameters": [
          {
            "in": "query",
            "name": "error",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login Page"
      },
      "post": {
        "operationId": "login_submit_portal_login_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_submit_portal_login_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login Submit"
      }
    },
...
    "/portal/signup": {
      "get": {
        "operationId": "signup_page_portal_signup_get",
        "parameters": [
          {
            "in": "query",
            "name": "error",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Signup Page"
      },
      "post": {
        "operationId": "signup_submit_portal_signup_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_signup_submit_portal_signup_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Signup Submit"
      }
    },
    "/portal/verify-email": {
      "get": {
        "operationId": "verify_email_page_portal_verify_email_get",
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Email Page"
      }
    }
  }
}